Introduce AttributeFilterBase component and customization
JIRA: RAIL-4323
